clt-chatbot-system/01.workspaces/apps/chatbot-admin-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
SQLAlchemy==2.0.27
psycopg[binary]==3.1.19
python-multipart==0.0.8
